+++ src/pages/helper/data.json
@@ -1,0 +1,182 @@
+[
+  {
+      "id": 1,
+      "username": "witchy_coder",
+      "title": "Spooky Sorting Algorithms",
+      "tags": ["algorithms", "sorting", "Halloween Challenge"],
+      "created_at": "2024-10-20T00:00:00Z",
+      "votes": 5,
+      "views": 300
+  },
+  {
+      "id": 2,
+      "username": "ghostly_dev",
+      "title": "Haunted Binary Tree Traversal",
+      "tags": ["trees", "binary tree", "Halloween Challenge"],
+      "created_at": "2024-10-19T19:00:00Z",
+      "votes": 2,
+      "views": 150
+  },
+  {
+      "id": 3,
+      "username": "frankenstein_code",
+      "title": "Zombie Apocalypse Survival Simulation",
+      "tags": ["simulation", "survival", "Halloween Challenge"],
+      "created_at": "2024-10-18T15:30:00Z",
+      "votes": 8,
+      "views": 450
+  },
+  {
+      "id": 4,
+      "username": "batman_developer",
+      "title": "Witch's Brew: Recipe API",
+      "tags": ["API", "recipe", "Halloween Challenge"],
+      "created_at": "2024-10-17T11:00:00Z",
+      "votes": 1,
+      "views": 200
+  },
+  {
+      "id": 5,
+      "username": "pumpkin_master",
+      "title": "Creepy Code Review",
+      "tags": ["review", "coding", "Halloween Challenge"],
+      "created_at": "2024-10-16T09:15:00Z",
+      "votes": 3,
+      "views": 350
+  },
+  {
+      "id": 6,
+      "username": "halloween_hacker",
+      "title": "Trick or Treat: A Coding Adventure",
+      "tags": ["adventure", "gaming", "Halloween Challenge"],
+      "created_at": "2024-10-15T14:00:00Z",
+      "votes": 7,
+      "views": 220
+  },
+  {
+      "id": 7,
+      "username": "zombie_coder",
+      "title": "AI in Zombie Defense",
+      "tags": ["AI", "zombies", "Halloween Challenge"],
+      "created_at": "2024-10-14T09:45:00Z",
+      "votes": 10,
+      "views": 500
+  },
+  {
+      "id": 8,
+      "username": "pumpkin_spice",
+      "title": "Spiced Up Algorithms",
+      "tags": ["algorithms", "spice", "Halloween Challenge"],
+      "created_at": "2024-10-13T12:00:00Z",
+      "votes": 4,
+      "views": 180
+  },
+  {
+      "id": 9,
+      "username": "skeleton_developer",
+      "title": "Skeletons in the Closet: Refactoring",
+      "tags": ["refactoring", "skeleton", "Halloween Challenge"],
+      "created_at": "2024-10-12T08:30:00Z",
+      "votes": 6,
+      "views": 270
+  },
+  {
+      "id": 10,
+      "username": "vampire_dev",
+      "title": "Bite-Sized Python Tips",
+      "tags": ["python", "tips", "Halloween Challenge"],
+      "created_at": "2024-10-11T10:15:00Z",
+      "votes": 12,
+      "views": 600
+  },
+  {
+      "id": 11,
+      "username": "mummy_coder",
+      "title": "Unwrapping the Mystery of JavaScript",
+      "tags": ["javascript", "mystery", "Halloween Challenge"],
+      "created_at": "2024-10-10T15:45:00Z",
+      "votes": 5,
+      "views": 320
+  },
+  {
+      "id": 12,
+      "username": "goblin_dev",
+      "title": "Goblin's Guide to SQL",
+      "tags": ["SQL", "database", "Halloween Challenge"],
+      "created_at": "2024-10-09T11:00:00Z",
+      "votes": 3,
+      "views": 210
+  },
+  {
+      "id": 13,
+      "username": "werewolf_programmer",
+      "title": "Howl at the Moon: Web Dev Essentials",
+      "tags": ["web development", "essentials", "Halloween Challenge"],
+      "created_at": "2024-10-08T17:30:00Z",
+      "votes": 9,
+      "views": 400
+  },
+  {
+      "id": 14,
+      "username": "cursed_dev",
+      "title": "Cursed Code Practices to Avoid",
+      "tags": ["best practices", "cursed", "Halloween Challenge"],
+      "created_at": "2024-10-07T13:15:00Z",
+      "votes": 4,
+      "views": 190
+  },
+  {
+      "id": 15,
+      "username": "spooky_dev",
+      "title": "Ghosts of Past Projects: Lessons Learned",
+      "tags": ["lessons", "projects", "Halloween Challenge"],
+      "created_at": "2024-10-06T16:00:00Z",
+      "votes": 11,
+      "views": 450
+  },
+  {
+      "id": 16,
+      "username": "haunting_dev",
+      "title": "Haunting Your Code: Debugging Tips",
+      "tags": ["debugging", "tips", "Halloween Challenge"],
+      "created_at": "2024-10-05T14:30:00Z",
+      "votes": 2,
+      "views": 160
+  },
+  {
+      "id": 17,
+      "username": "vampire_hacker",
+      "title": "Nightmare Fuel: Security Best Practices",
+      "tags": ["security", "best practices", "Halloween Challenge"],
+      "created_at": "2024-10-04T18:00:00Z",
+      "votes": 15,
+      "views": 700
+  },
+  {
+      "id": 18,
+      "username": "phantom_programmer",
+      "title": "Phantom Frameworks: What's Hot?",
+      "tags": ["frameworks", "trends", "Halloween Challenge"],
+      "created_at": "2024-10-03T20:00:00Z",
+      "votes": 1,
+      "views": 120
+  },
+  {
+      "id": 19,
+      "username": "creepy_coder",
+      "title": "Creepy Crawlies in Your Codebase",
+      "tags": ["codebase", "bugs", "Halloween Challenge"],
+      "created_at": "2024-10-02T21:15:00Z",
+      "votes": 7,
+      "views": 280
+  },
+  {
+      "id": 20,
+      "username": "zombie_hacker",
+      "title": "Post-Mortem: Evaluating Your Work",
+      "tags": ["evaluation", "work", "Halloween Challenge"],
+      "created_at": "2024-10-01T23:00:00Z",
+      "votes": 6,
+      "views": 330
+  }
+]
