+++ src/style/pages/helper.scss
@@ -13,7 +13,6 @@ $button-margin: 0 0.5em; // Add margin for spacing
     background-color: $background-color;
     color: $text-color;
     padding: 1em;
-    border-radius: 0.5em;
 
     h2 {
         text-align: center;
@@ -59,7 +58,7 @@ $button-margin: 0 0.5em; // Add margin for spacing
         display: flex;
         flex-direction: column;
         align-items: center;
-        gap: 1em;
+        gap: 0.75em;
         margin: 1em 0;
 
         &__item {
